Change game exclusion logic and make Game ID a long - If a user is in a game and tries to join quickmatch or a lobby, redirect them to the running game - If a user is in a lobby and tries to join quickmatch or a lobby, remove them from the original lobby - If a user is in quickmatch queue and tries to join a lobby remove them from quickmatch queue - The long data type for Game IDs significantly reduces the probability that a randomly generated ID will have previously been generated as another Game ID

+++ src/main/java/com/skyegibney/finar/game/events/FinarGameOverEvent.java
@@ -1,4 +1,6 @@
 package com.skyegibney.finar.game.events;
 
+import java.util.UUID;
+
 public record FinarGameOverEvent(
-    int gameId, String player1, String player2, String result, String winner, int[] winningMoves) {}
+        long gameId, String player1, String player2, String result, String winner, int[] winningMoves) {}
