+++ src/main/java/com/skyegibney/finar/game/events/TimeUpdateEvent.java
@@ -1,3 +1,5 @@
 package com.skyegibney.finar.game.events;
 
-public record TimeUpdateEvent(int gameId, long player1Time, long player2Time) {}
+import java.util.UUID;
+
+public record TimeUpdateEvent(long gameId, long player1Time, long player2Time) {}
